apparently interpreting an unsigned int as signed leads to bugs... plus fixed some off-by-one errors.

+++ _fragmentseeker.pxd
@@ -9,7 +9,7 @@ cdef extern from "bit.h":
 	int BITSLOT(int b)
 	int SETBIT(ULong a[], int b)
 	int CLEARBIT(ULong a[], int b)
-	int TESTBIT(ULong a[], int b)
+	ULong TESTBIT(ULong a[], int b)
 	int BITNSLOTS(int nb)
 	int IDX(int i, int j, int jmax, int kmax)
 	int abitcount(ULong vec[], int slots)
